requirements.txt
discord.py
requests
asyncio
os
